zoho-ai-sdr-unified/requirements.txt
zcatalyst-sdk==1.0.3
requests==2.32.3
typing-extensions==4.12.2